+++ lib/zakuro/calculation/range/transfer/gengou_scroller.rb
@@ -38,18 +38,32 @@ module Zakuro
               year.months.each_with_index do |month, index|
                 scroll.run(month: month)
                 gengou = scroll.to_gengou
-                year.months[index] = Monthly::Month.new(
-                  context: month.context,
-                  month_label: month.month_label,
-                  first_day: Monthly::FirstDay.new(
-                    remainder: month.first_day.remainder,
-                    average_remainder: month.first_day.average_remainder,
-                    western_date: gengou.start_date.clone
-                  ),
-                  solar_terms: month.solar_terms, gengou: gengou,
-                  meta: month.meta
+                year.months[index] = replace_gengou(
+                  month: month, gengou: gengou
                 )
               end
+            end
+
+            #
+            # 月情報のうち元号を差し替える
+            #
+            # @param [Monthly::Month] month 月情報
+            # @param [Base::Gengou] gengou 元号
+            #
+            # @return [Monthly::Month] 月情報
+            #
+            def replace_gengou(month:, gengou:)
+              Monthly::Month.new(
+                context: month.context,
+                month_label: month.month_label,
+                first_day: Monthly::FirstDay.new(
+                  remainder: month.first_day.remainder,
+                  average_remainder: month.first_day.average_remainder,
+                  western_date: gengou.start_date.clone
+                ),
+                solar_terms: month.solar_terms, gengou: gengou,
+                meta: month.meta
+              )
             end
           end
         end
